requirements.txt
# Core Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# API Integrations
requests>=2.31.0        # For API calls (weather, memes, etc.)
aiohttp>=3.9.0          # Async HTTP requests

# Date/Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Music Features
yt-dlp>=2023.11.16      # YouTube music playback (replacement for youtube-dl)
PyNaCl>=1.5.0           # Audio support for discord.py

# AI Features (optional - for AI responses)
openai>=1.3.0           # For OpenAI GPT integration (optional)

# Utilities
asyncio-throttle>=1.0.2 # Rate limiting